requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]

# Data Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Azure Storage for CSV files
azure-storage-blob==12.19.0
azure-identity==1.15.0

# CSV and Data Processing
pandas==2.1.4
python-csv==0.0.13

# HTTP Client for testing and external calls
httpx==0.25.2
requests==2.31.0

# Date and Time handling
python-dateutil==2.8.2

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development and Debugging
python-dotenv==1.0.0
typing-extensions==4.8.0

# Logging and Monitoring
structlog==23.2.0

# CORS support
fastapi-cors==0.0.6

# File handling and utilities
aiofiles==23.2.1

# UUID generation
uuid==1.30

# Concurrent operations
asyncio==3.4.3
